test(registry): profiles and comparisons
